+++ CotEditor/Sources/Shortcut.swift
@@ -151,13 +151,12 @@ struct Shortcut {
             charactersIgnoringModifiers.count == 1
         else { return nil }
         
-        // correct Backspace and Forward Delete keys
+        // correct Backspace key
         //  -> Backspace:      The key above the Return.
         //     Forward Delete: The key with printed "Delete" where next to the ten key pad.
         // cf. https://developer.apple.com/documentation/appkit/nsmenuitem/1514842-keyequivalent
         let keyEquivalent: String = switch event.specialKey {
-            case NSEvent.SpecialKey.delete:        String(NSEvent.SpecialKey.backspace.unicodeScalar)
-            case NSEvent.SpecialKey.deleteForward: String(NSEvent.SpecialKey.delete.unicodeScalar)
+            case NSEvent.SpecialKey.delete: String(NSEvent.SpecialKey.backspace.unicodeScalar)
             default: charactersIgnoringModifiers
         }
         
@@ -266,7 +265,8 @@ struct Shortcut {
         .downArrow: "↓",
         .leftArrow: "←",
         .rightArrow: "→",
-        .delete: "⌦",
+        .deleteForward: "⌦",
+        .delete: "⌫",
         .backspace: "⌫",
         .home: "↖",
         .end: "↘",